refactoring and creating public-api file
